refactor: remove redux implementation for HeaderBar. No longer needed with self-sufficient RefreshButton component
