+++ app/containers/Settings/Settings.jsx
@@ -9,7 +9,7 @@ import { recoverWallet } from '../../modules/generateWallet'
 import Panel from '../../components/Panel'
 import SelectInput from '../../components/Inputs/SelectInput'
 import StyledReactSelect from '../../components/Inputs/StyledReactSelect/StyledReactSelect'
-import HeaderBar from '../../components/HeaderBar'
+import HeaderBar from '../../components/HeaderBar/HeaderBar'
 import SettingsItem from '../../components/Settings/SettingsItem'
 import SettingsLink from '../../components/Settings/SettingsLink'
 import WalletRecoveryPanel from '../../components/Settings/WalletRecoveryPanel'
